--- runelite-client/src/main/java/net/runelite/client/plugins/freezetimersv2/HeadIcons.java
@@ -1,12 +1,0 @@
-package net.runelite.client.plugins.freezetimersv2;
-
-public class HeadIcons
-{
-	public static final int MELEE = 0;
-	public static final int RANGED = 1;
-	public static final int MAGIC = 2;
-	public static final int RETRIBUTION = 3;
-	public static final int SMITE = 4;
-	public static final int REDEMPTION = 5;
-	public static final int RANGE_MAGE = 6;
-}
