+++ runelite-client/src/main/resources/net/runelite/client/plugins/effecttimers/teleblockimmune.png
